in-15 Add different policies for admin and user

+++ investment-app/TRINV.IdentityServer/TRINV.IdentityServer/TRINV.IdentityServer/Data/Seed/PredefinedUsers.cs
@@ -22,7 +22,7 @@ internal static class PredefinedUsers
                 AccountEnabled = true,
                 Claims =
                 {
-                    new Claim(Claims.RoleKey, ((int)Role.Admin).ToString()),
+                    new Claim(Claims.RoleKey, Role.Admin.ToString()),
                 }
             }
         };
